requirements.txt
absl-py==0.8.1
amqp==2.1.4
ansible==2.4.2.0
appdirs==1.4.3
asgiref==3.2.3
asn1crypto==0.24.0
astor==0.7.1
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.4
beautifulsoup4==4.8.1
billiard==3.5.0.3
bleach==3.1.0
bs4==0.0.1
celery==4.1.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
cssselect==1.1.0
cycler==0.10.0
decorator==4.1.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==1.11
django-bootstrap3==9.1.0
django-crispy-forms==1.8.1
django-filter==1.1.0
django-formtools==2.1
django-import-export==2.0.1
django-ranged-response==0.2.0
django-redis-cache==1.7.1
django-rest-swagger==2.1.2
django-reversion==3.0.5
django-simple-captcha==0.5.6
django-suit==0.2.25
django-uuslug==1.2.0
django-widget-tweaks==1.4.5
djangorestframework==3.7.3
djangorestframework-bulk==0.2.1
dominate==2.4.0
ecdsa==0.13
elasticsearch==7.5.1
elasticsearch-dsl==7.1.0
entrypoints==0.3
enum-compat==0.0.2
et-xmlfile==1.0.1
eventlet==0.21.0
fake-useragent==0.1.11
flake8==3.7.9
Flask==0.11.1
Flask-Bootstrap==3.3.7.0
ForgeryPy==0.1
future==0.18.2
gast==0.3.2
google-pasta==0.1.8
greenlet==0.4.12
grpcio==1.23.0
gunicorn==19.6.0
h2==2.6.2
h5py==2.10.0
hpack==3.0.0
httplib2==0.17.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.6
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==0.24
itypes==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
kombu==4.0.2
ldap3==2.4
lxml==4.4.1
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.0
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
mkl-service==2.3.0
more-itertools==7.2.0
mysqlclient==1.3.12
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.3
odfpy==1.4.1
olefile==0.44
openapi-codec==1.3.2
opencv-python==4.1.2.30
openpyxl==3.0.3
pandocfilters==1.4.2
paramiko==2.4.0
parse==1.12.1
parso==0.5.1
passlib==1.7.1
pickleshare==0.7.5
Pillow==4.3.0
pipenv==2018.11.26
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.1
pyasn1==0.4.2
pycodestyle==2.5.0
pycparser==2.18
pyee==6.0.0
PyExecJS==1.5.1
pyflakes==2.1.1
Pygments==2.4.2
pymongo==3.9.0
PyNaCl==1.2.1
pyotp==2.3.0
pyparsing==2.4.5
pyppeteer==0.0.25
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtWebEngine==5.12.1
pyquery==1.4.1
pyreadline==2.1
pyrsistent==0.15.5
python-dateutil==2.8.1
python-slugify==4.0.0
pytz==2017.3
pywin32==225
pywinpty==0.5.5
PyYAML==3.12
pyzmq==18.1.1
qtconsole==4.6.0
redis==2.10.6
requests==2.18.4
requests-html==0.10.0
scipy==1.3.1
Send2Trash==1.5.0
simplejson==3.13.2
six==1.11.0
soupsieve==1.9.5
spyder-kernels==1.8.0
sqlparse==0.3.0
tablib==0.14.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
torch==1.3.1
torchvision==0.4.2
tornado==6.0.3
tqdm==4.38.0
traitlets==4.3.3
ufzh==1.0.1
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
virtualenv==16.7.9
virtualenv-clone==0.5.3
visitor==0.1.3
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
websockets==8.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.7
xlwt==1.3.0
zipp==0.6.0
